fix: don't flag "can not" to join as a "compound noun"
"Cannot" is valid word, but "can not" is also valid and, more importantly, it's not a compound noun.

It is marginally a noun though as in "cans and cannots", which is why it was detected.

Now it will only flag words when joined they will be a noun but won't also be a verb.