fix: use community maintained fork of tower-lsp
`tower-lsp-server` is a community fork of `tower-lsp`, which appears to
be unmaintained.

This fixes https://github.com/Automattic/harper/issues/139 (thanks to
https://github.com/tower-lsp-community/tower-lsp-server/commit/5ac6bceb6d6283c3b7d101b32ca25a0f0a4c0b0e,
which landed in `tower-lsp-server` v0.21.0).

Unfortunately, this wasn't a drop in replacement. `tower-lsp-server` has
made some changes, and also depends on a completely different url/uri
library. Here are the relevant changes:

- `tower-lsp-server`: https://github.com/tower-lsp-community/tower-lsp-server/blob/v0.21.1/CHANGELOG.md
- `lsp-types`'s url/uri change: https://github.com/gluon-lang/lsp-types/commit/e3d0ed287e85bc453d7f77a7c9e6d39620b8d1ef