feat: mass nouns after a/an etc.
Work-in-progress

Looks for mass nouns being treated as singular or plural nouns.

- group and sort `sequence_expr` `gen_then_from_is` etc.
- Added a new `agreement` `TokenKind` and added it to two other linters.
- Adding singular/plural & countable/mass flags to dictionary.

Does not yet handle adjectives between the determiner/quantifier and the noun.

Does not yet handle the case where the mass noun is part of a compound noun and therefor not an error.

Does not yet offer specific synonym such as "a tip" for "an advice".