chore: mark comparatives and superlatives in dictionary
These are the ones whose last letter is doubled since the affix logic doesn't handle that.

Add comments to `affixes.json` linking the annotation flags for the base form comparative and superlative with the deirved term ones.

Add `get_degree` to `WordMetadata` and as a bonus also `get_person` for pronouns.

* Positive degree is not yet queryable as I suspect there could be false positives.