fix: canonicalize workspace_root in configuration
If the LSP client does not provide any mechanism to determine the root
of the workspace, we need to resolve things relative to the current
directory, but previously we tried that as `.` which it turns out
`.try_resolve_in()` does not like.

This changes the behaviour to always `.canonicalize()` the root of the
workspace so that subsequent path resolution starts from an absolute
path.

Fixes #1707

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>