feat(gitcommit): use tree-sitter for gitcommit
The previous gitcommit parser naively took the contents until the
first comment. If there are inline comments, such as when squashing two
commits in an interactive rebase, the text after those comments would
not be checked.

Instead, we use tree-sitter-gitcommit parsing and mask the nodes that
are subject, message lines or breaking change descriptions. As before,
we re-parse these parts with the existing markdown parser.

A new crate `harper-git-commit` is added.