Compromise to implement the dialect enum
Now DictWordMetadata takes a DialectFlagsEnum instead of a
EnglishDialect. There are some adaptations that needed to be made, for
example creating the Dialect trait, and adding to this trait a function
that gets most used dialects from document using a provided language
(LanguageFamily, which is a variant of the enum that doesn't have the
dialect contained inside it). This  function should not be used for any
of the specific DialectFlags implementations, only for the
DialectFlagsEnum.

On the counterpart, the variant of the function that doesn't receive any
languages should only be used by the specific implementations, which
should be progressively more private, being replaced in general by the
enum.

By the way, the original idea was using in the DictWordMetadata struct a
Box<dyn DialogFlag>, but in the end the trait is not dyn compatible,
hence this workaround